requirements.txt
ttkthemes==2.4.0
Pillow==6.2.1
